refactor(torghut): extract subscription persistence
